+++ jax/_src/numpy/lax_numpy.py
@@ -3427,12 +3427,24 @@ def _normalize_index(index, axis_size):
     lax.add(index, axis_size_val),
     index)
 
-@_wraps(np.take_along_axis, update_doc=False)
-@partial(jit, static_argnames=('axis',))
-def take_along_axis(arr, indices, axis: Optional[int]):
+
+TAKE_ALONG_AXIS_DOC = """
+Unlike :func:`numpy.take_along_axis`, :func:`jax.numpy.take_along_axis` takes
+an optional ``mode`` parameter controlling how out-of-bounds indices should be
+handled. By default, out-of-bounds indices are clamped into range. In a future
+change, out-of-bounds indices will return invalid (e.g., ``NaN``) values
+instead. See :attr:`jax.numpy.ndarray.at` for more discussion
+of out-of-bounds indexing in JAX.
+"""
+
+@_wraps(np.take_along_axis, update_doc=False,
+        lax_description=TAKE_ALONG_AXIS_DOC)
+@partial(jit, static_argnames=('axis', 'mode'))
+def take_along_axis(arr, indices, axis: Optional[int],
+                    mode: Optional[Union[str, lax.GatherScatterMode]] = None):
   _check_arraylike("take_along_axis", arr, indices)
   # index_dtype = dtypes.dtype(indices)
-  # TODO(phawkins): reenalbe this check after fixing callers
+  # TODO(phawkins): reenable this check after fixing callers
   # if not dtypes.issubdtype(index_dtype, integer):
   #   raise TypeError("take_along_axis indices must be of integer type, got "
   #                   f"{str(index_dtype)}")
@@ -3510,7 +3522,8 @@ def take_along_axis(arr, indices, axis: Optional[int]):
     collapsed_slice_dims=tuple(collapsed_slice_dims),
     start_index_map=tuple(start_index_map))
   # TODO(phawkins): change the mode to "fill".
-  return lax.gather(arr, gather_indices, dnums, tuple(slice_sizes))
+  return lax.gather(arr, gather_indices, dnums, tuple(slice_sizes),
+                    mode="clip" if mode is None else mode)
 
 ### Indexing